+++ src/game/Game.h
@@ -5,6 +5,7 @@
 #include "game/GameTypes.h"
 #include "game/GameConstants.h"
 #include "game/Physics.h"
+#include "game/Rendering.h"
 
 
 void UpdateGameObjects(Time time, GameObjectManager* gameObjectManager);
@@ -19,6 +20,3 @@ void InitMesh(Mesh* mesh);
 void UpdateGameObject(Time time, GameObject* gameObject); // called every frame update
 
 void UpdateTransform(Time time, Transform* transform);
-void UpdateMesh(Time time, Mesh* mesh);
-void DrawGizmos(Time time, Vector3 maxSize);
-void DrawBoundingBox(BoudingBox box);
